Fixed issue where packets spam
Added a feature to drop packets after detecting same destination and source

changed "of.OFPP_ALL" to "of.OFPP_FLOOD"
